removed unwanted files and refined the code
