+++ src/Scenes/Artists/Artists.tsx
@@ -1,10 +1,7 @@
-import { Flex, Text } from "palette"
 import React from "react"
+import { Text } from "palette"
+import { Tabs } from "react-native-collapsible-tab-view"
 
-export const ArtistsScreen = () => {
-  return (
-    <Flex flex={1} justifyContent="center" textAlign="center">
-      <Text>Artists Screen</Text>
-    </Flex>
-  )
+export const Artists = () => {
+  return <Tabs.FlatList data={[1, 2, 3, 4, 5]} renderItem={({ item }) => <Text>Artist {item}</Text>} />
 }
