+++ tsconfig.json
@@ -45,7 +45,7 @@
       "@routes/*": ["./routes/*"],
       "@helpers/*": ["./helpers/*"],
       "@assets/*": ["./assets/*"],
-    },                           
+    },
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
     // "types": [],                           /* Type declaration files to be included in compilation. */
